==> 01-scRNAseq Acute Wounds/05_subClustering/s4-Myeloid/s2_clean_immuneCells_mg_GOs.Rmd <==
---
title: "Marker genes and DE analysis res1"
author: "Zhuang Liu"
date: "02/12/2021 Updated: `r Sys.Date()`"
output:
  html_document:
    toc: true #add a table of contents (TOC)
    toc_depth: 3 #headers' level in the toc
    toc_float: true #toc visible even when scrolled
    theme: lumen
    highlight: tango
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, fig.align='center', fig.width=12, fig.height = 8, warning=FALSE, message=FALSE)
```

```{r warning=FALSE, message=FALSE}
library(Seurat)
library(SeuratObject)
library(SeuratDisk)
library(tidyverse)
library(magrittr)
library(ggplot2)
library(patchwork)
library(cowplot)
library(harmony)
```

####################################################
# Step. Marker genes identification
####################################################
## S. Known marker genes
```{r fig.width=9, fig.height=6}
hswound.immu.com <- readRDS("s1_clean_immu_myeloid.rds")
allMyeloidCells_clean_CellID <- colnames(hswound.immu.com)
saveRDS(allMyeloidCells_clean_CellID, file = "allMyeloidCells_clean_CellID.rds")

hswound.immu.com
DimPlot(hswound.immu.com, reduction = "umap", group.by='SCT_snn_res.1', label=TRUE, label.size = 6) + NoAxes()
```
```{r fig.height=8, fig.width=12}
plotcom <- FeaturePlot(
  hswound.immu.com,
  features=c('COL1A1', 'COL1A2', 'KRT5', 'KRT10', 'LYZ', 'HLA-DRA', 'CD68', 'CD79A', 'CD7', 'CD3D', 'CD207', 'PECAM1'), 
  combine=F, cols = c("grey","red")
)

# apply theme to each feature plot
for(i in 1:length(plotcom)){
  plotcom[[i]] <- plotcom[[i]] + NoLegend() + NoAxes()
}

cowplot::plot_grid(plotlist = plotcom, ncol = 4)
```

```{r fig.height=11}
plotcom <- FeaturePlot(
  hswound.immu.com,
  features=c('CD68', 'CD163', 'CD14', 'CD209', 'FCGR2A', 'MRC1', 'CD207', 'CD1A', 'CD1C', 'CD1B', 'ITGAX', 
             'CD83', 'LAMP3', 'THBD', 'CLEC9A', 'MKI67'), 
  combine=F, cols = c("grey","red")
)

# apply theme to each feature plot
for(i in 1:length(plotcom)){
  plotcom[[i]] <- plotcom[[i]] + NoLegend() + NoAxes()
}

cowplot::plot_grid(plotlist = plotcom, ncol = 4)
```

```{r}
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "orig.ident", label.size = 4.5) + NoAxes() 

table(hswound.immu.com$SCT_snn_res.0.5)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.0.5", label.size = 4.5) + NoAxes() 

table(hswound.immu.com$SCT_snn_res.0.8)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.0.8", label.size = 4.5) + NoAxes()

table(hswound.immu.com$SCT_snn_res.1)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1", label.size = 4.5) + NoAxes()

table(hswound.immu.com$SCT_snn_res.1.2)
DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "SCT_snn_res.1.2", label.size = 4.5) + NoAxes()

DimPlot(object = hswound.immu.com, 
        reduction = "umap", 
        label = TRUE, 
        group.by = "MainCellTypes", label.size = 5) + NoAxes() + ggtitle("")
```

```{r fig.height=4}
DimPlot(hswound.immu.com, reduction = "umap", pt.size = .001,group.by = "SCT_snn_res.1", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")
DimPlot(hswound.immu.com, reduction = "umap", pt.size = .001,group.by = "MainCellTypes", split.by = 'Condition', ncol = 4, label = F) + NoLegend()+ ggtitle("")
FeaturePlot(hswound.immu.com, features = "MKI67", split.by = 'Condition')
```

```{r}
doublet_scores <- data.table::fread("/Users/zhuliu/Desktop/scRNA_STseq/proj_10X_woundhealing/03_results/01-Seurat-PreAnalysis/02_Seurat_BatchCorrection/s1_Seurat_allSample_harmony/py_doublet_OriScores.txt")

doublet_scores_f <- colnames(hswound.immu.com) %>% as.data.frame() %>% rename("ID" = ".") %>% left_join(., doublet_scores, by=c("ID" = "barcodename"))
hswound.immu.com$doublet_scores <- doublet_scores_f$doublet_scores

# Determine metrics to plot present in hswound.krt.com@meta.data
metrics <-  c("nCount_RNA", "nFeature_RNA", "S.Score", "G2M.Score", "percent.mt", "doublet_scores")

p3 <- FeaturePlot(
  hswound.immu.com,
  features = metrics,
  cols = c("gray","red"), combine = FALSE)
p2_list <- lapply(p3, function(x) {x + NoAxes() + theme(text = element_text(size = 10))})
cowplot::plot_grid(plotlist = p2_list, ncol = 3)

VlnPlot(hswound.immu.com, features = metrics, group.by='SCT_snn_res.1', pt.size = 0, ncol = 2)

ggplot(hswound.immu.com@meta.data, aes(x = SCT_snn_res.1, fill = Condition, alpha = Patient)) +
    geom_bar(position = "fill")
```

```{r fig.width=6}
DefaultAssay(hswound.immu.com) <- "SCT"
hswound.immu.com@active.ident <- hswound.immu.com$SCT_snn_res.1
hswound.immu.com <- BuildClusterTree(object = hswound.immu.com, assay = "SCT", slot = "scale.data")
suppressMessages(library(scales))
cluster.color <- hue_pal()(length(levels(hswound.immu.com)))
# pull the tree
hswound.clustertree <- Tool(object = hswound.immu.com, slot = "BuildClusterTree")
# plot the tree
suppressMessages(require(ape))
{plot.phylo(x = hswound.clustertree, direction = "rightwards", use.edge.length = TRUE, font = 4, tip.color = cluster.color, edge.width = 2, label.offset = 5, no.margin = TRUE)
tiplabels(pch = 19, col = cluster.color, adj = 3, cex = 2)}
```

## S. Visualizing marker genes for each cluster
```{r fig.width=12, fig.height=30}
DefaultAssay(hswound.immu.com) <- "RNA"
markers <- data.table::fread("s1_clean_immu_myeloid_mg_res1.txt")

markers_top10 <- markers %>% group_by(cluster) %>% top_n(10, avg_log2FC)
fac_levs <- unique(markers$cluster)

plot_list <- list()
for (i in seq_along(fac_levs)) {
  #print(paste0("Cluster ", i, " :Top 10 marker genes"))
  names <- fac_levs[i]
  plot_list[[i]] <- DotPlot(hswound.immu.com, features = markers_top10[markers_top10$cluster == names, ]$gene, col.max = 5, col.min = 0, 
                                dot.scale = 4, dot.min = 0.1,
                                assay = "RNA", group.by = "SCT_snn_res.1") + RotatedAxis() + coord_flip() + xlab("Marker genes") + ylab("") +
    theme(axis.text.x=element_text(size=10), axis.text.y=element_text(size=10)) + ggtitle(label = names)
}

wrap_plots(plot_list,ncol = 2) #+ plot_layout(guides = 'collect') & theme(legend.position = 'right')
```

## S. featurePlot Visualizing marker genes for each cluster
```{r fig.height=14}
markers_top10 <- markers %>% group_by(cluster) %>% top_n(20, avg_log2FC)
fac_levs <- unique(markers$cluster)

for (i in seq_along(fac_levs)) {
  print(paste0("Plot cluster", i-1))
  names <- fac_levs[i]
  plotgenes= markers_top10[markers_top10$cluster == names, ]$gene
  plot_list <- FeaturePlot(
    hswound.immu.com,
    features=plotgenes,
    combine=FALSE, cols = c("gray","red")
  )
  # apply theme to each feature plot
  for(i in 1:length(plot_list)){
    plot_list[[i]] <- plot_list[[i]] + NoLegend() + NoAxes()
  }
  print(wrap_plots(plot_list, ncol = 5))
}
```

####################################################
# Step. Gene Ontology analysis for each cluster
####################################################
## S. GO results
```{r echo=FALSE, warning=FALSE, message=FALSE}
GO_BP <- function(i=NULL){
  datatmp <- markers %>% filter(cluster == i)
  require(clusterProfiler) #loading the package if it is not loaded
  require(org.Hs.eg.db)
  Enriched_gene_up <- as.character(unique(datatmp$entrezgene_id[1:200])[1:100])
  ego_BP_up <- enrichGO(gene = Enriched_gene_up,
                        OrgDb         = org.Hs.eg.db,
                        ont           = "BP",
                        keyType = "ENTREZID",
                        pAdjustMethod = "BH",
                        pvalueCutoff  = 0.5,
                        qvalueCutoff  = 0.2,
                        minGSSize = 10,
                        maxGSSize = 500,
                        readable = T)
   ego_BP_up_f <- as.data.frame(ego_BP_up)
   if (nrow(ego_BP_up_f) > 0) {
     dotplot(ego_BP_up, showCategory=10, label_format = 100) + 
       ggtitle(paste0("Cluster ", i, "-GO_BP")) + 
    theme(
      #panel.grid.major.x = element_blank(),
      #panel.grid.major.y = element_blank(),
      panel.grid.minor = element_blank(),
      panel.border = element_rect(size = 0.5, color = "black"),
      axis.text.x = element_text(size = 14, color = "black"),
      axis.text.y = element_text(size = 14, color = "black"))
    } else{
      print("No enriched terms")
    }
}
```

## `r paste0("Cluster ", 0)`
```{r fig.height=5, fig.width=12}
GO_BP(i=0)
```

## `r paste0("Cluster ", 1)`
```{r fig.height=5, fig.width=12}
GO_BP(i=1)
```

## `r paste0("Cluster ", 2)`
```{r fig.height=5, fig.width=12}
GO_BP(i=2)
```

## `r paste0("Cluster ", 3)`
```{r fig.height=5, fig.width=12}
GO_BP(i=3)
```

## `r paste0("Cluster ", 4)`
```{r fig.height=5, fig.width=12}
GO_BP(i=4)
```

## `r paste0("Cluster ", 5)`
```{r fig.height=5, fig.width=12}
GO_BP(i=5)
```

## `r paste0("Cluster ", 6)`
```{r fig.height=5, fig.width=12}
GO_BP(i=6)
```

## `r paste0("Cluster ", 7)`
```{r fig.height=5, fig.width=12}
GO_BP(i=7)
```

## `r paste0("Cluster ", 8)`
```{r fig.height=5, fig.width=12}
GO_BP(i=8)
```

## `r paste0("Cluster ", 9)`
```{r fig.height=5, fig.width=12}
GO_BP(i=9)
```

## `r paste0("Cluster ", 10)`
```{r fig.height=5, fig.width=12}
GO_BP(i=10)
```

## `r paste0("Cluster ", 11)`
```{r fig.height=5, fig.width=12}
GO_BP(i=11)
```

## `r paste0("Cluster ", 12)`
```{r fig.height=5, fig.width=12}
GO_BP(i=12)
```

# SessionInfo
```{r}
sessionInfo()
```
